requirements.txt
alembic==1.4.2
altair==4.1.0
appdirs==1.4.3
appnope==0.1.0
argon2-cffi==21.1.0
asgiref==3.2.3
astor==0.8.1
attrs==21.2.0
autopep8==1.5.4
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.1
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
cachelib==0.3.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
convertdate==2.3.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
Flask-Dance==3.1.0
Flask-JWT-Extended==3.25.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.15.1
Flask==1.1.2
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==1.6.0
ipykernel==6.5.0
ipython-genutils==0.2.0
ipython==7.29.0
ipywidgets==7.6.5
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.4
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.4.3
mistune==0.8.4
multitasking==0.0.9
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.4
oauthlib==3.1.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.3.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.6.0
pyarrow==6.0.0
pycodestyle==2.6.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
PyJWT==1.7.1
PyMeeus==0.5.11
Pympler==0.9
pyparsing==2.4.7
pyrsistent==0.18.0
pyTelegramBotAPI==4.0.0
python-dateutil==2.8.2
python-editor==1.0.4
python-http-client==3.3.2
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
Send2Trash==1.8.0
sendgrid==6.8.1
six==1.16.0
smmap==5.0.0
SQLAlchemy-Utils==0.36.8
SQLAlchemy==1.3.17
sqlparse==0.3.0
starkbank-ecdsa==1.1.1
streamlit==1.2.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
URLObject==2.4.3
validators==0.18.2
virtualenv==20.0.20
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
WTForm==1.0
WTForms==2.3.3
yfinance==0.1.64
zipp==3.1.0